requirements.txt
playwright>=1.42.0
pyyaml>=6.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
